+++ changelog.md
@@ -28,6 +28,8 @@
 - Added std/enumutils module containing `genEnumCaseStmt` macro that generates
   case statement to parse string to enum.
 
+- Added `freshIdentNodes` in macros, for replacing `nnkSym` nodes to identifier nodes.
+
 - Removed deprecated `iup` module from stdlib, it has already moved to
   [nimble](https://github.com/nim-lang/iup).